VTX setup refined

+++ libraries/AP_VideoTX/AP_SmartAudio.cpp
@@ -221,7 +221,7 @@ void AP_SmartAudio::update_vtx_params()
             if (_vtx_use_set_freq) {
                 set_frequency(vtx.get_configured_frequency_mhz(), false);
             } else {
-                set_channel(vtx.get_configured_band() * VTX_MAX_CHANNELS + vtx.get_configured_channel());
+                set_channel(static_cast<uint16_t>(vtx.get_configured_band()) * VTX_MAX_CHANNELS + vtx.get_configured_channel());
             }
         } else if (_vtx_power_change_pending) {
             debug("update power (ver %u)", _protocol_version);
